+++ src/equalizer.js
@@ -1,30 +1,15 @@
 import fs from 'fs'
 import path from 'path'
 import { ERRORS } from './constants'
+import fileReader from './fileReader'
+import { diffArrays } from 'diff'
 
 export default function equalize({
   languages,
   localesDirectory,
   referenceLocale,
 }) {
-  const termsPerLanguage = {}
-
-  walk(localesDirectory, languages, filepath => {
-    const language = path.basename(filepath, '.js').replace('.json', '')
-
-    const terms = require(filepath)
-
-    Object.keys(terms).forEach(term => {
-      if (!termsPerLanguage[language]) {
-        termsPerLanguage[language] = {
-          filepath,
-        }
-      }
-      termsPerLanguage[language][term] = {
-        referenceValue: terms[term],
-      }
-    })
-  })
+  const termsPerLanguage = fileReader({ languages, localesDirectory })
 
   const hasEmptyLanguage = languages.some(
     language =>
@@ -34,56 +19,79 @@ export default function equalize({
 
   if (hasEmptyLanguage) {
     return {
-      data: languages.find(
-        language =>
-          !termsPerLanguage[language] ||
-          Object.keys(termsPerLanguage[language]).length === 0
-      ),
-      error: ERRORS.ERROR_NO_KEYS_LOCALE,
+      error: {
+        code: ERRORS.ERROR_NO_KEYS_LOCALE,
+        data: languages.find(
+          language =>
+            !termsPerLanguage[language] ||
+            Object.keys(termsPerLanguage[language]).length === 0
+        ),
+      },
     }
   }
 
   const processedLanguages = {}
 
   Object.keys(termsPerLanguage[referenceLocale]).forEach(key => {
+    const referenceOrder = Object.keys(termsPerLanguage[referenceLocale])
+
     languages.forEach(language => {
       if (!processedLanguages[language]) {
         processedLanguages[language] = {
           missingKeys: [],
+          wrongOrderKeys: [],
         }
       }
 
       if (!termsPerLanguage[language][key]) {
-        processedLanguages[language].filepath = termsPerLanguage[
-          referenceLocale
-        ].filepath
-          .replace(`${referenceLocale}.js`, `${language}.js`)
-          .replace(process.cwd(), '')
-
         processedLanguages[language].missingKeys.push({
           key,
           ...termsPerLanguage[referenceLocale][key],
         })
       }
+
+      const languageOrder = Object.keys(termsPerLanguage[language])
+
+      processedLanguages[language].wrongOrderKeys = getCorrectLines(
+        referenceOrder,
+        languageOrder
+      )
     })
   })
 
   return processedLanguages
 }
 
-function walk(dir, languages, callback) {
-  const files = fs.readdirSync(dir)
-  files.forEach(file => {
-    var filepath = path.join(dir, file)
-    const stats = fs.statSync(filepath)
-    if (stats.isDirectory()) {
-      walk(filepath, callback)
-    } else if (
-      stats.isFile() &&
-      (languages.includes(path.basename(filepath, '.js')) ||
-        languages.includes(path.basename(filepath, '.json')))
-    ) {
-      callback(filepath, stats)
+function getCorrectLines(referenceOrder, languageOrder) {
+  const diff = diffArrays(referenceOrder, languageOrder)
+
+  const correctLines = diff.reduce((chunks, line) => {
+    const key = line.value[0]
+    if (line.removed || line.added) {
+      if (!chunks[key]) {
+        chunks[key] = {}
+      }
+    }
+
+    if (line.removed) {
+      chunks[key] = {
+        ...chunks[key],
+        wrongLine: languageOrder.indexOf(key),
+      }
+    } else if (line.added) {
+      chunks[key] = {
+        ...chunks[key],
+        correctLine: referenceOrder.indexOf(key),
+      }
+    }
+
+    return chunks
+  }, {})
+
+  return Object.keys(correctLines).map(key => {
+    return {
+      key,
+      ...correctLines[key],
     }
   })
 }