+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.1.0] - 2019-02-15
+
+### Added
+
+- Create `wrongOrderKeys` table
+
+### Changed
+
+- `table` to `missingKeysTable`
+
 ## [2.0.1] - 2018-12-19
 
 - Update npm package build script